@@ -1,18 +1,17 @@
 <script lang="ts">
   import { onMount } from "svelte";
-  import { API } from "@/api";
-  import { error } from "@/mainStore";
+  import { API } from "./ts/api.ts";
+  import { appState } from "ts/state.ts";
   import {goto} from "@mateothegreat/svelte5-router";
 
-  export let email: string;
-  export let token: string;
+  let {email, token} = $props<{email: string; name: string;}>();
 
   onMount(async () => {
     try {
       await API.user.confirmGitEmail(email, token);
       goto("/user/settings");
     } catch (e) {
-      $error = e;
+      appState.setError(e);
       goto("/user/search");
     }
   });
