@@ -1,48 +1,30 @@
 import {defineConfig} from '@rsbuild/core';
 import {pluginSvelte} from '@rsbuild/plugin-svelte';
-import path from 'path';
+import * as process from "node:process";
+import {pluginCssMinimizer} from "@rsbuild/plugin-css-minimizer";
 
 export default defineConfig({
     environments: {
         // Configure the web environment for browsers
         web: {
             plugins: [
-                pluginSvelte()
+                pluginSvelte(),
+                pluginCssMinimizer()
             ],
             source: {
                 entry: {
-                    index: './src/index-client.ts', //creates a html
-                },
-                alias: {
-                    '/images': path.resolve(__dirname, 'public/images')
+                    index: './src/index.ts',
                 }
             },
             output: {
-                //assetPrefix: './',
-                target: 'web'
-            }
-        },
-        // Configure the node environment for SSR
-        ssr: {
-            plugins: [
-                pluginSvelte({
-                  svelteLoaderOptions: {
-                    compilerOptions: {
-                      //@ts-ignore -> this is the right option
-                      generate: 'server'
-                    }
-                  }
-                })
-            ],
-            source: {
-                entry: {
-                    server: './src/index-server.ts', //creates a js
-                }
-            },
-            output: {
-                // Use 'node' target for the Node.js outputs
-                target: 'node'
+                target: 'web',
+                minify:  process.env.NODE_ENV === 'production',
             }
         }
     },
+    dev: {
+        hmr: false,
+        liveReload: true,
+    }
+    //TODO npx run stage fails with root: './src'. This makes no sense
 });
