@@ -1,6 +1,5 @@
 <script lang="ts">
   import queryString from "query-string";
-  import { lastEthRoute } from "@/ethStore";
   import {route} from "@mateothegreat/svelte5-router";
 
   export let location;
@@ -16,7 +15,7 @@
       Your MetaMask browser plugin is currently connected to the network with ID {queryParams.actual}.
       However, FlatFeeStack requires to use {queryParams.required}. Please
       change your network in MetaMask and
-      <a use:route href="{$lastEthRoute}">click here</a> to return to the previous page.
+      <a use:route href="">click here</a> to return to the previous page.
     </p>
   </div>
 </div>
